+++ src/conf/rpc.rs
@@ -6,22 +6,37 @@ use serde::{Deserialize, Serialize};
 use smart_default::SmartDefault;
 
 /// RPC connection settings.
-#[derive(Clone, Debug, Deserialize, Serialize, SmartDefault)]
+#[derive(Clone, Copy, Debug, Deserialize, Serialize, SmartDefault)]
 #[serde(default)]
 pub struct Rpc {
-    /// Duration, after which remote RPC client will be considered idle if no
-    /// heartbeat messages received. Defaults to `10s`.
+    /// Duration, after which remote RPC client will be considered idle
+    /// if no heartbeat messages received.
+    ///
+    /// It applies to all related pipelines as default value, but can be
+    /// overridden for each specific case via Control API.
+    ///
+    /// Defaults to `10s`.
     #[default(Duration::from_secs(10))]
     #[serde(with = "humantime_serde")]
     pub idle_timeout: Duration,
 
     /// Duration, after which the server deletes the client session if
     /// the remote RPC client does not reconnect after it is idle.
+    ///
+    /// It applies to all related pipelines as default value, but can be
+    /// overridden for each specific case via Control API.
+    ///
+    /// Defaults to `10s`.
     #[default(Duration::from_secs(10))]
     #[serde(with = "humantime_serde")]
     pub reconnect_timeout: Duration,
 
     /// Interval of sending `Ping`s from the server to the client.
+    ///
+    /// It applies to all related pipelines as default value, but can be
+    /// overridden for each specific case via Control API.
+    ///
+    /// Defaults to `3s`.
     #[default(Duration::from_secs(3))]
     #[serde(with = "humantime_serde")]
     pub ping_interval: Duration,