+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All user visible changes to this project will be documented in this file. This p
         - `on_leave` ([#63]);
         - `on_start` ([#91]);
         - `on_stop` ([#91]).
+    - Configuration of `Member`'s Client API RPC settings ([#95]).
 - Signalling:
     - Dynamic `Peer`s creation when client connects ([#28]);
     - Auto-removing `Peer`s when `Member` disconnects ([#28]);
@@ -66,8 +67,9 @@ All user visible changes to this project will be documented in this file. This p
 [#81]: /../../pull/81
 [#84]: /../../pull/84
 [#86]: /../../pull/86
-[#94]: /../../pull/94
 [#91]: /../../pull/91
+[#94]: /../../pull/94
+[#95]: /../../pull/95